debug

+++ src/dataset.py
@@ -33,10 +33,10 @@ class Dataset:
 
     @tf.function
     def __augment(self, img, mask):
-        if tf.random.uniform((), dtype=tf.dtypes.float64) > 0.5:
+        if tf.random.uniform((), dtype=tf.dtypes.float32) > 0.5:
             img = tf.image.flip_left_right(img)
             mask = tf.image.flip_left_right(mask)
-        if tf.random.uniform((), dtype=tf.dtypes.float64) > 0.5:
+        if tf.random.uniform((), dtype=tf.dtypes.float32) > 0.5:
             img = tf.image.central_crop(img, central_fraction=0.8)
             img = tf.image.resize(img, self.image_shape)
             mask = tf.image.central_crop(mask, central_fraction=0.8)
